device: Move LED management to main.c
The contents of device.h are currently a mix of application- and
FOTA-related declarations; it would be better for them to be one or
the other. Arbitrarily, let's choose to keep the FOTA-related
declarations, so that moving the app contents elsewhere lets us move
these files to lib/.

Start with the LED declarations, which are application-specific.

Signed-off-by: Marti Bolivar <marti.bolivar@linaro.org>