conf: add 1 extra buffer for MQTT use
Give the MQTT subsys an extra buffer to use incase the 1 default
buffer is allocated for an extra long period of time.  This avoids
an odd condition where a buffer will be lost during the initial
ICMP / ARP broadcasts and MQTT will fail until the device is
rebooted.

NOTE: This may not be needed in the future as the MQTT subsys
gets more love.

Signed-off-by: Michael Scott <michael.scott@linaro.org>